+++ src/layer/riscv/convolution_packnto1_fp16s.h
@@ -140,14 +140,14 @@ static void convolution_packnto1_fp16sa_rvv(const Mat& bottom_blob, Mat& top_blo
         {
             for (int j = 0; j < outw; j++)
             {
-                __fp16 sum = 0.f;
+                __fp16 sum = (__fp16)0.f;
 
                 if (bias_data_ptr)
                 {
                     sum = bias_data_ptr[p];
                 }
 
-                vfloat16m1_t _sum = __riscv_vfmv_v_f_f16m1(0.f, vl);
+                vfloat16m1_t _sum = __riscv_vfmv_v_f_f16m1((__fp16)0.f, vl);
 
                 const __fp16* kptr = weight_data_fp16.channel(p);
