+++ src/layer/riscv/convolution_winograd_transform_packn_fp16s.h
@@ -91,7 +91,7 @@ static void conv3x3s1_winograd63_transform_input_packn_fp16sa_rvv(const Mat& bot
                     __riscv_vse16_v_f16m1(tmp[3][m], _tmp3m, vl);
                     __riscv_vse16_v_f16m1(tmp[4][m], _tmp4m, vl);
 
-                    vfloat16m1_t _tmp56a = __riscv_vfmacc_vf_f16m1(_r06, 4.f, __riscv_vfmacc_vf_f16m1(_r02, (__fp16)-1.25f, _r04, vl), vl);
+                    vfloat16m1_t _tmp56a = __riscv_vfmacc_vf_f16m1(_r06, (__fp16)4.f, __riscv_vfmacc_vf_f16m1(_r02, (__fp16)-1.25f, _r04, vl), vl);
                     vfloat16m1_t _tmp56b = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, (__fp16)2.f, vl), (__fp16)-2.5f, _r03, vl), (__fp16)0.5f, _r05, vl);
 
                     vfloat16m1_t _tmp5m = __riscv_vfadd_vv_f16m1(_tmp56a, _tmp56b, vl);
@@ -131,7 +131,7 @@ static void conv3x3s1_winograd63_transform_input_packn_fp16sa_rvv(const Mat& bot
                     vfloat16m1_t _r0tm1 = __riscv_vfadd_vv_f16m1(_tmp12a, _tmp12b, vl);
                     vfloat16m1_t _r0tm2 = __riscv_vfsub_vv_f16m1(_tmp12a, _tmp12b, vl);
 
-                    vfloat16m1_t _tmp34a = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_tmp06, 0.25f, _tmp02, vl), (__fp16)-1.25f, _tmp04, vl);
+                    vfloat16m1_t _tmp34a = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_tmp06, (__fp16)0.25f, _tmp02, vl), (__fp16)-1.25f, _tmp04, vl);
                     vfloat16m1_t _tmp34b = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp01, (__fp16)0.5f, vl), (__fp16)-2.5f, _tmp03, vl), (__fp16)2.f, _tmp05, vl);
 
                     vfloat16m1_t _r0tm3 = __riscv_vfadd_vv_f16m1(_tmp34a, _tmp34b, vl);
@@ -253,7 +253,7 @@ static void conv3x3s1_winograd63_transform_output_packn_fp16sa_rvv(const Mat& to
 
                     vfloat16m1_t _tmp1m = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_tmp135a, (__fp16)2.f, _tmp135b, vl), (__fp16)16.f, _tmp135c, vl);
                     vfloat16m1_t _tmp3m = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_tmp135a, (__fp16)8.f, _tmp135b, vl), (__fp16)4.f, _tmp135c, vl);
-                    vfloat16m1_t _tmp5m = __riscv_vfadd_vv_f16m1(__riscv_vfadd_vv_f16m1(_out0tm7, _tmp135a, vl), __riscv_vfmacc_vf_f16m1(_tmp135c, 32.f, _tmp135b, vl), vl);
+                    vfloat16m1_t _tmp5m = __riscv_vfadd_vv_f16m1(__riscv_vfadd_vv_f16m1(_out0tm7, _tmp135a, vl), __riscv_vfmacc_vf_f16m1(_tmp135c, (__fp16)32.f, _tmp135b, vl), vl);
                     __riscv_vse16_v_f16m1(tmp[1][m], _tmp1m, vl);
                     __riscv_vse16_v_f16m1(tmp[3][m], _tmp3m, vl);
                     __riscv_vse16_v_f16m1(tmp[5][m], _tmp5m, vl);
@@ -366,17 +366,17 @@ static void conv3x3s1_winograd43_transform_input_packn_fp16sa_rvv(const Mat& bot
                     vfloat16m1_t _r04 = __riscv_vle16_v_f16m1(r0 + packn * 4, vl);
                     vfloat16m1_t _r05 = __riscv_vle16_v_f16m1(r0 + packn * 5, vl);
 
-                    vfloat16m1_t _tmp01a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, sq2, vl), -sq2_d2, _r03, vl);
-                    vfloat16m1_t _tmp01b = __riscv_vfmacc_vf_f16m1(_r04, -2.f, _r02, vl);
-                    vfloat16m1_t _tmp23a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, sq2_d2, vl), -sq2, _r03, vl);
-                    vfloat16m1_t _tmp23b = __riscv_vfmacc_vf_f16m1(_r04, -0.5f, _r02, vl);
+                    vfloat16m1_t _tmp01a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, (__fp16)sq2, vl), (__fp16)-sq2_d2, _r03, vl);
+                    vfloat16m1_t _tmp01b = __riscv_vfmacc_vf_f16m1(_r04, (__fp16)-2.f, _r02, vl);
+                    vfloat16m1_t _tmp23a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, (__fp16)sq2_d2, vl), (__fp16)-sq2, _r03, vl);
+                    vfloat16m1_t _tmp23b = __riscv_vfmacc_vf_f16m1(_r04, (__fp16)-0.5f, _r02, vl);
 
-                    vfloat16m1_t _tmp0m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r00, _r04, vl), -2.5f, _r02, vl);
+                    vfloat16m1_t _tmp0m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r00, _r04, vl), (__fp16)-2.5f, _r02, vl);
                     vfloat16m1_t _tmp1m = __riscv_vfsub_vv_f16m1(_tmp01b, _tmp01a, vl);
                     vfloat16m1_t _tmp2m = __riscv_vfadd_vv_f16m1(_tmp01b, _tmp01a, vl);
                     vfloat16m1_t _tmp3m = __riscv_vfsub_vv_f16m1(_tmp23b, _tmp23a, vl);
                     vfloat16m1_t _tmp4m = __riscv_vfadd_vv_f16m1(_tmp23b, _tmp23a, vl);
-                    vfloat16m1_t _tmp5m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r01, _r05, vl), -2.5f, _r03, vl);
+                    vfloat16m1_t _tmp5m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r01, _r05, vl), (__fp16)-2.5f, _r03, vl);
 
                     __riscv_vse16_v_f16m1(tmp[0][m], _tmp0m, vl);
                     __riscv_vse16_v_f16m1(tmp[1][m], _tmp1m, vl);
@@ -404,17 +404,17 @@ static void conv3x3s1_winograd43_transform_input_packn_fp16sa_rvv(const Mat& bot
                     vfloat16m1_t _r04 = __riscv_vle16_v_f16m1(tmp[m][4], vl);
                     vfloat16m1_t _r05 = __riscv_vle16_v_f16m1(tmp[m][5], vl);
 
-                    vfloat16m1_t _tmp01a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, sq2, vl), -sq2_d2, _r03, vl);
-                    vfloat16m1_t _tmp01b = __riscv_vfmacc_vf_f16m1(_r04, -2.f, _r02, vl);
-                    vfloat16m1_t _tmp23a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, sq2_d2, vl), -sq2, _r03, vl);
-                    vfloat16m1_t _tmp23b = __riscv_vfmacc_vf_f16m1(_r04, -0.5f, _r02, vl);
+                    vfloat16m1_t _tmp01a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, (__fp16)sq2, vl), (__fp16)-sq2_d2, _r03, vl);
+                    vfloat16m1_t _tmp01b = __riscv_vfmacc_vf_f16m1(_r04, (__fp16)-2.f, _r02, vl);
+                    vfloat16m1_t _tmp23a = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_r01, (__fp16)sq2_d2, vl), (__fp16)-sq2, _r03, vl);
+                    vfloat16m1_t _tmp23b = __riscv_vfmacc_vf_f16m1(_r04, (__fp16)-0.5f, _r02, vl);
 
-                    vfloat16m1_t _tmp0m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r00, _r04, vl), -2.5f, _r02, vl);
+                    vfloat16m1_t _tmp0m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r00, _r04, vl), (__fp16)-2.5f, _r02, vl);
                     vfloat16m1_t _tmp1m = __riscv_vfsub_vv_f16m1(_tmp01b, _tmp01a, vl);
                     vfloat16m1_t _tmp2m = __riscv_vfadd_vv_f16m1(_tmp01b, _tmp01a, vl);
                     vfloat16m1_t _tmp3m = __riscv_vfsub_vv_f16m1(_tmp23b, _tmp23a, vl);
                     vfloat16m1_t _tmp4m = __riscv_vfadd_vv_f16m1(_tmp23b, _tmp23a, vl);
-                    vfloat16m1_t _tmp5m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r01, _r05, vl), -2.5f, _r03, vl);
+                    vfloat16m1_t _tmp5m = __riscv_vfmacc_vf_f16m1(__riscv_vfadd_vv_f16m1(_r01, _r05, vl), (__fp16)-2.5f, _r03, vl);
 
                     __riscv_vse16_v_f16m1(r0_tm_0, _tmp0m, vl);
                     __riscv_vse16_v_f16m1(r0_tm_1, _tmp1m, vl);
@@ -473,7 +473,7 @@ static void conv3x3s1_winograd43_transform_output_packn_fp16sa_rvv(const Mat& to
         const Mat out0_tm = top_blob_tm.channel(p);
         Mat out0 = top_blob.channel(p);
 
-        vfloat16m1_t _bias0 = biasptr ? __riscv_vle16_v_f16m1(biasptr + p * packn, vl) : __riscv_vfmv_v_f_f16m1(0.f, vl);
+        vfloat16m1_t _bias0 = biasptr ? __riscv_vle16_v_f16m1(biasptr + p * packn, vl) : __riscv_vfmv_v_f_f16m1((__fp16)0.f, vl);
 
         // NOTE variable length array
         __fp16 tmp[4][6][packn];
@@ -507,9 +507,9 @@ static void conv3x3s1_winograd43_transform_output_packn_fp16sa_rvv(const Mat& to
                     vfloat16m1_t _tmp13b = __riscv_vfsub_vv_f16m1(_r03, _r04, vl);
 
                     vfloat16m1_t _tmp0m = __riscv_vfadd_vv_f16m1(__riscv_vfadd_vv_f16m1(_r00, _tmp02a, vl), _tmp02b, vl);
-                    vfloat16m1_t _tmp1m = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp13a, sq2_d2, vl), sq2, _tmp13b, vl);
-                    vfloat16m1_t _tmp2m = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp02a, 0.5f, vl), 2.f, _tmp02b, vl);
-                    vfloat16m1_t _tmp3m = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_r05, sq2_d4, _tmp13a, vl), sq2_m2, _tmp13b, vl);
+                    vfloat16m1_t _tmp1m = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp13a, (__fp16)sq2_d2, vl), (__fp16)sq2, _tmp13b, vl);
+                    vfloat16m1_t _tmp2m = __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp02a, (__fp16)0.5f, vl), (__fp16)2.f, _tmp02b, vl);
+                    vfloat16m1_t _tmp3m = __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_r05, (__fp16)sq2_d4, _tmp13a, vl), (__fp16)sq2_m2, _tmp13b, vl);
 
                     __riscv_vse16_v_f16m1(tmp[0][m], _tmp0m, vl);
                     __riscv_vse16_v_f16m1(tmp[1][m], _tmp1m, vl);
@@ -539,9 +539,9 @@ static void conv3x3s1_winograd43_transform_output_packn_fp16sa_rvv(const Mat& to
                     vfloat16m1_t _tmp13b = __riscv_vfsub_vv_f16m1(_r03, _r04, vl);
 
                     vfloat16m1_t _out00 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfadd_vv_f16m1(__riscv_vfadd_vv_f16m1(_r00, _tmp02a, vl), _tmp02b, vl), vl);
-                    vfloat16m1_t _out01 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp13a, sq2_d2, vl), sq2, _tmp13b, vl), vl);
-                    vfloat16m1_t _out02 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp02a, 0.5f, vl), 2.f, _tmp02b, vl), vl);
-                    vfloat16m1_t _out03 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_r05, sq2_d4, _tmp13a, vl), sq2_m2, _tmp13b, vl), vl);
+                    vfloat16m1_t _out01 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp13a, (__fp16)sq2_d2, vl), (__fp16)sq2, _tmp13b, vl), vl);
+                    vfloat16m1_t _out02 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfmacc_vf_f16m1(__riscv_vfmul_vf_f16m1(_tmp02a, (__fp16)0.5f, vl), (__fp16)2.f, _tmp02b, vl), vl);
+                    vfloat16m1_t _out03 = __riscv_vfadd_vv_f16m1(_bias0, __riscv_vfmacc_vf_f16m1(__riscv_vfmacc_vf_f16m1(_r05, (__fp16)sq2_d4, _tmp13a, vl), (__fp16)sq2_m2, _tmp13b, vl), vl);
 
                     __riscv_vse16_v_f16m1(output0, _out00, vl);
                     __riscv_vse16_v_f16m1(output0 + packn, _out01, vl);
@@ -677,7 +677,7 @@ static void conv3x3s1_winograd23_transform_output_packn_fp16sa_rvv(const Mat& to
         const Mat out0_tm = top_blob_tm.channel(p);
         Mat out0 = top_blob.channel(p);
 
-        vfloat16m1_t _bias0 = biasptr ? __riscv_vle16_v_f16m1(biasptr + p * packn, vl) : __riscv_vfmv_v_f_f16m1(0.f, vl);
+        vfloat16m1_t _bias0 = biasptr ? __riscv_vle16_v_f16m1(biasptr + p * packn, vl) : __riscv_vfmv_v_f_f16m1((__fp16)0.f, vl);
 
         // NOTE variable length array
         __fp16 tmp[2][4][packn];
